added empId field in both retail and textile product class

+++ inventory-module/src/main/java/com/otsi/retail/inventory/service/ProductItemService.java
@@ -17,7 +17,7 @@ public interface ProductItemService {
 
 	List<ProductItemVo> getAllBarcodes(ProductItemVo vo);
 
-	ProductItemVo getBarcodeId(Long barcodeId);
+	ProductItemVo getBarcodeId(int barcodeId);
 
 	String updateProduct(ProductItemVo vo);
